Modularize integration tests APIs..
into the following modules:
- lib: with general purpose functions (e.g.: starters)
- sniffer
- template_provider
