+++ roles/tests-integration/tests/translator_integration.rs
@@ -1,7 +1,5 @@
-mod common;
-
-use common::sniffer::MessageDirection;
 use const_sv2::{MESSAGE_TYPE_SETUP_CONNECTION, MESSAGE_TYPE_SUBMIT_SHARES_EXTENDED};
+use integration_tests_sv2::{sniffer::*, *};
 use roles_logic_sv2::parsers::{CommonMessages, Mining, PoolMessages};
 
 // This test runs an sv2 translator between an sv1 mining device and a pool. the connection between
@@ -10,12 +8,12 @@ use roles_logic_sv2::parsers::{CommonMessages, Mining, PoolMessages};
 // shares.
 #[tokio::test]
 async fn translation_proxy() {
-    let (_tp, tp_addr) = common::start_template_provider(None).await;
-    let (_pool, pool_addr) = common::start_pool(Some(tp_addr)).await;
+    let (_tp, tp_addr) = start_template_provider(None).await;
+    let (_pool, pool_addr) = start_pool(Some(tp_addr)).await;
     let (pool_translator_sniffer, pool_translator_sniffer_addr) =
-        common::start_sniffer("0".to_string(), pool_addr, false, None).await;
-    let (_, tproxy_addr) = common::start_sv2_translator(pool_translator_sniffer_addr).await;
-    let _mining_device = common::start_mining_device_sv1(tproxy_addr).await;
+        start_sniffer("0".to_string(), pool_addr, false, None).await;
+    let (_, tproxy_addr) = start_sv2_translator(pool_translator_sniffer_addr).await;
+    let _mining_device = start_mining_device_sv1(tproxy_addr).await;
     pool_translator_sniffer
         .wait_for_message_type(MessageDirection::ToUpstream, MESSAGE_TYPE_SETUP_CONNECTION)
         .await;
